requirements.txt
Django==3.0.8
django_filters==0.2.1
djangorestframework==3.11.0
localflavor==1.8
pendulum==2.1.0
Pillow==7.2.0
selenium==3.141.0
